Filter sidebarPanes by object key
