+++ shared/ui/Authentication/actions.ts
@@ -53,7 +53,7 @@ import { moveCursorToLine } from "../Stream/api-functions";
 import { localStore } from "../utilities/storage";
 import { emptyObject, uuid } from "../utils";
 import { HostApi } from "../webview-api";
-import { WebviewPanels } from "@codestream/webview/ipc/webview.protocol.common";
+import { WebviewPanels } from "@codestream/protocols/api";
 
 export enum SignupType {
 	JoinTeam = "joinTeam",
@@ -207,7 +207,7 @@ export const authenticate =
 							userId: response.extra.userId,
 							eligibleJoinCompanies: response.extra.eligibleJoinCompanies,
 							accountIsConnected: response.extra.accountIsConnected,
-						}),
+						})
 					);
 				case LoginResult.NotOnTeam:
 					return dispatch(
@@ -217,7 +217,7 @@ export const authenticate =
 							// login is for resuming previous sessions and this error means you haven't ever fully signed into the extension
 							email: (params as PasswordLoginParams).email,
 							token: response.extra.token,
-						}),
+						})
 					);
 				default:
 					throw response.error;
@@ -243,7 +243,7 @@ export const generateLoginCode =
 						username: "",
 						password: "",
 					},
-				}),
+				})
 			);
 		} else {
 			throw response.status;
@@ -271,7 +271,7 @@ export const onLogin =
 					bootstrapCore,
 				};
 			},
-			"bootstrap",
+			"bootstrap"
 		);
 
 		await dispatch(
@@ -290,7 +290,7 @@ export const onLogin =
 					isFirstPageview,
 					currentCodemarkId: response.state.codemarkId,
 				},
-			}),
+			})
 		);
 
 		const { context } = getState();
@@ -330,7 +330,7 @@ export const completeSignup =
 			provider?: string;
 			byDomain?: boolean;
 			setEnvironment?: { environment: string; serverUrl: string };
-		},
+		}
 	) =>
 	async (dispatch, getState: () => CodeStreamState) => {
 		const tokenUrl =
@@ -372,7 +372,7 @@ export const completeSignup =
 					prefPath: ["sidebarPanes", WebviewPanels.CICD, "removed"],
 					value: true,
 				},
-			]),
+			])
 		);
 
 		const providerName = extra.provider
@@ -395,7 +395,7 @@ export const completeAcceptInvite =
 			provider?: string;
 			byDomain?: boolean;
 			setEnvironment?: { environment: string; serverUrl: string };
-		},
+		}
 	) =>
 	async (dispatch, getState: () => CodeStreamState) => {
 		const tokenUrl =
@@ -439,7 +439,7 @@ export const completeAcceptInvite =
 					prefPath: ["sidebarPanes", WebviewPanels.CICD, "removed"],
 					value: true,
 				},
-			]),
+			])
 		);
 
 		const providerName = extra.provider
@@ -495,7 +495,7 @@ export const validateSignup =
 							accountIsConnected: response.extra && response.extra.accountIsConnected,
 							isWebmail: response.extra.isWebmail,
 							provider,
-						}),
+						})
 					);
 				case LoginResult.NotOnTeam:
 					HostApi.instance.track("Account Created", {
@@ -508,7 +508,7 @@ export const validateSignup =
 							email: response.extra && response.extra.email,
 							token: response.extra && response.extra.token,
 							provider,
-						}),
+						})
 					);
 				case LoginResult.ProviderConnectFailed:
 				// @ts-ignore - reset the otc and cascade to the default case
