+++ shared/ui/Stream/GettingStarted.tsx
@@ -8,7 +8,8 @@ import { CodeStreamState } from "../store";
 import Tooltip from "./Tooltip";
 import { ComposeArea } from "./ReviewNav";
 import { openPanel, openModal } from "../store/context/actions";
-import { WebviewPanels, WebviewModals } from "@codestream/webview/ipc/webview.protocol.common";
+import { WebviewModals } from "@codestream/webview/ipc/webview.protocol.common";
+import { WebviewPanels } from "@codestream/protocols/api";
 import { isConnected } from "../store/providers/reducer";
 import { setUserPreference } from "./actions";
 import { Button } from "../src/components/Button";
