+++ shared/ui/Stream/OpenPullRequests.tsx
@@ -22,7 +22,7 @@ import styled from "styled-components";
 import { PullRequestDetailsLoading } from "@codestream/webview/Stream/PullRequestLoading";
 
 import { OpenUrlRequestType, ReviewCloseDiffRequestType } from "@codestream/protocols/webview";
-import { WebviewPanels } from "@codestream/webview/ipc/webview.protocol.common";
+import { WebviewPanels } from "@codestream/protocols/api";
 import { setPaneMaximized } from "@codestream/webview/Stream/actions";
 import { useAppDispatch, useAppSelector } from "@codestream/webview/utilities/hooks";
 import { disposePoll, fluctuatePoll } from "@codestream/webview/utils";
@@ -249,7 +249,7 @@ export const OpenPullRequests = React.memo((props: Props) => {
 		const prSupportedProviders = providerSelectors
 			.getSupportedPullRequestHosts(state)
 			.filter(
-				provider => !teamSettings?.limitCodeHost || teamSettings?.codeHostProviders?.[provider.id],
+				provider => !teamSettings?.limitCodeHost || teamSettings?.codeHostProviders?.[provider.id]
 			);
 		const prConnectedProviders = providerSelectors.getConnectedSupportedPullRequestHosts(state);
 		const prConnectedProvidersWithErrors = prConnectedProviders.filter(_ => _.hasAccessTokenError);
@@ -320,7 +320,7 @@ export const OpenPullRequests = React.memo((props: Props) => {
 			...state,
 			[action.provider]: action.error,
 		}),
-		{},
+		{}
 	);
 	const [prCommitsRange, setPrCommitsRange] = React.useState<string[]>([]);
 	const [openRepos, setOpenRepos] = React.useState<ReposScmPlusName[]>([]);
@@ -340,7 +340,7 @@ export const OpenPullRequests = React.memo((props: Props) => {
 				[action.provider]: action.isLoading,
 			};
 		},
-		{},
+		{}
 	);
 	const [isLoadingPRGroup, setIsLoadingPRGroup] = React.useState<number | undefined>(undefined);
 	const [individualLoadingPR, setIndividualLoadingPR] = React.useState("");
@@ -349,15 +349,15 @@ export const OpenPullRequests = React.memo((props: Props) => {
 		{ providerId: string; index: number } | undefined
 	>(undefined);
 	const previousPRConnectedProvidersWithErrorsCount = usePrevious<number>(
-		derivedState.PRConnectedProvidersWithErrorsCount,
+		derivedState.PRConnectedProvidersWithErrorsCount
 	);
 	const previousPRConnectedProvidersCount = usePrevious<number>(
-		derivedState.PRConnectedProvidersCount,
+		derivedState.PRConnectedProvidersCount
 	);
 
 	const saveQueries = (providerId, queries) => {
 		dispatch(
-			setUserPreference({ prefPath: ["pullRequestQueries", providerId], value: [...queries] }),
+			setUserPreference({ prefPath: ["pullRequestQueries", providerId], value: [...queries] })
 		);
 	};
 
@@ -365,7 +365,7 @@ export const OpenPullRequests = React.memo((props: Props) => {
 		async (
 			theQueries: FetchProviderDefaultPullResponse,
 			options?: { force?: boolean; alreadyLoading?: boolean },
-			src: string | undefined = undefined,
+			src: string | undefined = undefined
 		) => {
 			let count: number | undefined = undefined;
 			let activePrListedCount = 0;
@@ -387,7 +387,7 @@ export const OpenPullRequests = React.memo((props: Props) => {
 					}
 
 					activePrListedIndex = queriesByProvider?.findIndex(
-						_ => _?.name === "Waiting on my Review",
+						_ => _?.name === "Waiting on my Review"
 					);
 					// console.warn("Loading the PRs... in the loop", queryStrings);
 					try {
@@ -406,7 +406,7 @@ export const OpenPullRequests = React.memo((props: Props) => {
 							const twoWeekAgoTimestamp = +new Date(Date.now() - 12096e5);
 							if (activePrListedIndex >= 0) {
 								activePrListedCount = response[activePrListedIndex].filter(
-									activePr => activePr.createdAt > twoWeekAgoTimestamp,
+									activePr => activePr.createdAt > twoWeekAgoTimestamp
 								).length;
 							}
 							// Previously called setPullRequestGroups(updatedPullRequestGroups);
@@ -454,7 +454,7 @@ export const OpenPullRequests = React.memo((props: Props) => {
 			PRConnectedProviders,
 			derivedState.allRepos,
 			derivedState.myPullRequests,
-		],
+		]
 	);
 
 	useEffect(() => {
@@ -479,8 +479,8 @@ export const OpenPullRequests = React.memo((props: Props) => {
 		};
 	}, [queries]);
 
-	/* 
-   TODO Fix this effect disposing / recreating every run of fetchPRs 
+	/*
+   TODO Fix this effect disposing / recreating every run of fetchPRs
    received notification codestream/didChangeData from host on user object
    which triggers PRConnectedProviders / getConnectedSupportedPullRequestHosts selector
    with no changed data
@@ -515,7 +515,7 @@ export const OpenPullRequests = React.memo((props: Props) => {
 		(async () => {
 			const defaultQueriesResponse: FetchProviderDefaultPullResponse = await HostApi.instance.send(
 				FetchProviderDefaultPullRequestsType,
-				{},
+				{}
 			);
 			if (defaultQueriesResponse) {
 				// Update default queries for users in a non-destructive way
@@ -601,7 +601,7 @@ export const OpenPullRequests = React.memo((props: Props) => {
 					throwOnError: true,
 					test: false,
 					index: index,
-				}),
+				})
 			).unwrap();
 			// No longer calling set`PullRequestGroups - updatePullRequestFilter action already does this
 		} catch (ex) {
@@ -649,7 +649,7 @@ export const OpenPullRequests = React.memo((props: Props) => {
 			setUserPreference({
 				prefPath: ["pullRequestProviderHidden", providerId],
 				value: !pullRequestProviderHidden[providerId],
-			}),
+			})
 		);
 	};
 
@@ -685,7 +685,7 @@ export const OpenPullRequests = React.memo((props: Props) => {
 			openPullRequestByUrl({
 				url,
 				options: { providerId, groupIndex: "-1" },
-			}),
+			})
 		)) as {
 			error?: string;
 		};
@@ -712,7 +712,7 @@ export const OpenPullRequests = React.memo((props: Props) => {
 		) {
 			fetchOnePR(
 				derivedState.currentPullRequestFromContext?.providerId,
-				derivedState.currentPullRequestFromContext?.id,
+				derivedState.currentPullRequestFromContext?.id
 			);
 		}
 	}, [derivedState.expandedPullRequestGroupIndex, derivedState.currentPullRequestFromContext]);
@@ -729,7 +729,7 @@ export const OpenPullRequests = React.memo((props: Props) => {
 		const myPullRequests = derivedState.myPullRequests;
 		if (myPullRequests) {
 			Object.values(myPullRequests)?.forEach(group =>
-				group.forEach(list => (total += list.length)),
+				group.forEach(list => (total += list.length))
 			);
 		}
 		return total;
@@ -1433,7 +1433,7 @@ export const OpenPullRequests = React.memo((props: Props) => {
 					setUserPreference({
 						prefPath: ["pullRequestQueryShowAllRepos"],
 						value: !derivedState.allRepos,
-					}),
+					})
 				),
 		},
 		{
@@ -1445,7 +1445,7 @@ export const OpenPullRequests = React.memo((props: Props) => {
 					setUserPreference({
 						prefPath: ["pullRequestQueryHideDescriptions"],
 						value: !derivedState.hideDescriptions,
-					}),
+					})
 				),
 		},
 		{
@@ -1457,7 +1457,7 @@ export const OpenPullRequests = React.memo((props: Props) => {
 					setUserPreference({
 						prefPath: ["pullRequestQueryHideLabels"],
 						value: !derivedState.hideLabels,
-					}),
+					})
 				),
 		},
 		// Not using this for now
@@ -1765,7 +1765,7 @@ export const OpenPullRequests = React.memo((props: Props) => {
 	function patchQueries(
 		defaultQueries: FetchProviderDefaultPullResponse,
 		pullRequestQueries: FetchProviderDefaultPullResponse,
-		saveQueries: (providerId, queries) => void,
+		saveQueries: (providerId, queries) => void
 	) {
 		Object.keys(pullRequestQueries).forEach(provider => {
 			// Little shimmy to update old default filters to our new syntax
@@ -1812,7 +1812,7 @@ export const OpenPullRequests = React.memo((props: Props) => {
 						query.query === "is:pr is:open involves:@me"
 					) {
 						const replacementQuery = defaultQueries[provider].find(
-							_ => _.name === "Waiting on my Review",
+							_ => _.name === "Waiting on my Review"
 						)?.query;
 						if (replacementQuery) {
 							pullRequestQueries![provider][index].query = replacementQuery;
