+++ shared/ui/index.tsx
@@ -72,7 +72,7 @@ import {
 	WebviewDidInitializeNotificationType,
 	ViewAnomalyNotificationType,
 } from "./ipc/webview.protocol";
-import { WebviewPanels } from "@codestream/webview/ipc/webview.protocol.common";
+import { WebviewPanels } from "@codestream/protocols/api";
 import { store } from "./store";
 import { bootstrap, reset } from "./store/actions";
 import {
