+++ shared/ui/ipc/webview.protocol.common.ts
@@ -1,5 +1,5 @@
 import { EnvironmentHost, GetFileScmInfoResponse } from "@codestream/protocols/agent";
-import { CSEligibleJoinCompany } from "@codestream/protocols/api";
+import { CSEligibleJoinCompany, WebviewPanels } from "@codestream/protocols/api";
 import { Position, Range } from "vscode-languageserver-types";
 
 import { NewPullRequestBranch } from "./webview.protocol";
@@ -35,45 +35,6 @@ export interface EditorSidebarLocation {
 export interface EditorSelection extends Range {
 	// MUST be identical to Range.end
 	cursor: Position;
-}
-
-/* NOTE: there can be dynamic panel names that begin with configure-provider- or configure-enterprise- */
-export enum WebviewPanels {
-	Activity = "activity",
-	Codemarks = "knowledge",
-	CodemarksForFile = "codemarks-for-file",
-	Export = "export",
-	FilterSearch = "filter-search",
-	Invite = "invite",
-	NewComment = "new-comment",
-	NewIssue = "new-issue",
-	NewReview = "new-review",
-	Profile = "profile", // DO NOT USE
-	PullRequest = "pull-request",
-	Review = "review",
-	Status = "status",
-	Tasks = "tasks",
-	LandingRedirect = "landing-redirect", // new customers and signins go here
-	PRInfo = "pr-info",
-	Integrations = "integrations",
-	GettingStarted = "gtting-started", // this is a typo but now baked into user data, so let's just leave it
-	NewPullRequest = "new-pull-request",
-	Flow = "flow",
-	Tester = "tester",
-	Sidebar = "sidebar",
-	OpenReviews = "open-reviews",
-	OpenPullRequests = "open-pull-requests",
-	Onboard = "onboard",
-	OnboardNewRelic = "onboard-newrelic",
-	Commits = "commits",
-	ErrorsInbox = "errorsinbox",
-	CodeError = "code-error",
-	Observability = "observability",
-	PixieDynamicLogging = "pixie-dynamic-logging",
-	MethodLevelTelemetry = "method-level-telemetry",
-	ObservabilityAnomaly = "observability-anomaly",
-	CICD = "ci-cd",
-	CodeAnalyzers = "code-analyzers",
 }
 
 // this is for mixpanel and maps the values from WebviewPanels to their
