+++ shared/ui/store/context/reducer.ts
@@ -1,7 +1,7 @@
 import { ActionType } from "../common";
 import * as preferencesActions from "../preferences/actions";
 // import * as sessionActions from "../session/actions";
-import { WebviewPanels } from "@codestream/webview/ipc/webview.protocol.common";
+import { WebviewPanels } from "@codestream/protocols/api";
 import { CodeStreamState } from "..";
 import { PreferencesActionsType } from "../preferences/types";
 import * as actions from "./actions";
